requirements.txt
google-cloud-storage>=2.10.0
pytest>=7.0.0
numpy